refactor(Day4): move into single file
